[interferon] Improve updates and dry-run mode
Improve dry-run mode to track updates and deletions which are part
of the difference with respect to Datadog. It now looks at the diff of
the message section and keep track of updates seperately from new
alerts. It also keeps track of deletions separately from dry-run clean
up. Also added more statistics around what is successfully created
versus what was intended to be created.